Basic Camera Movement
Added some of the basic camera movement such as left, right, forward, backward, zoom in and out and faster movement when holding control. The movement could be changed to react with the surface of the landscape, using a line trace to check the distance from the ground and using the old movement of SetActorLocation.