Building system has been added with basic placement checks and resource usage
The building system has been fully implemented with some changes in the future to improve it.
It checks for any building before placing and uses and checks for resources before placement and when placed.
The data has been organized inside a data table located in the Data folder.
The HUD has its own widget to add more widgets onto it without any problems.